Add support for Float16 in Vulkan kernels and improve numerical stability
- Implement zero-variance guard in layer normalization backward kernels for Float16 and Float64 to prevent gradient overflow.
- Update nonzero count kernel to use a more robust check for non-zero values.
- Enhance Vulkan backend to support grouped convolutions by adding a groups parameter in convolution backward functions.
- Introduce new Vulkan kernels for bilinear and nearest neighbor interpolation in Float16.
- Modify MaskRCNN and YOLO models to ensure proper data types and device compatibility.
- Optimize GELU backward computation by using a fused kernel to avoid overflow in Float16.
- Ensure statistics are correctly handled for backward passes in normalization layers, maintaining device compatibility.

+++ src/models/mask_rcnn.cpp
@@ -581,7 +581,7 @@ auto MaskRCNN::forward_train(const Variable& images,
                         "bilinear",
                         false
                     );
-                    resized_mask = resized_mask.squeeze(0).squeeze(0).to(Device::cpu());
+                    resized_mask = resized_mask.squeeze(0).squeeze(0).to(DType::Float32).to(Device::cpu());
 
                     // Copy to sampled_masks_cpu
                     auto target_mask_cpu = tenzor::select(sampled_masks_cpu, 0, i);
